Tier 2: drive `leaderboard()` from an ordered metric registry (#362)
* Tier 2: drive leaderboard() from an ordered metric registry

Replace the ~75-line procedural if-chain in `leaderboard()` with a declarative,
ordered registry of metric producers (`_LEADERBOARD_METRICS`). Each metric is a
small `_LeaderboardMetric` (key + producer + requires_baseline/always_on), and the
method body is now a single loop over the registry plus the existing sort/drop
post-processing. Output is byte-identical (column content and order).

Adds a composable `metrics=` argument: pass an iterable of metric keys
(e.g. ["elo", "winrate", "mrr"]) instead of the ~11 `include_*` booleans. When given
it overrides the flags; otherwise the flags map onto registry keys via
`_LEADERBOARD_FLAG_TO_KEY`, so the public signature stays fully backward compatible
(callers keep passing `include_*` / `elo_kwargs` / ... unchanged). Adding a new
column is now one registry entry, with no new parameter or branch.

The greedy-removal `MetricSpec` system is intentionally left untouched: it has a
different output shape (per-method scalar over arbitrary subsets) and is a separate
concern.

Behavior is pinned by tst/test_evaluator.py: the 19 existing tests pass unchanged,
plus 5 new registry tests including an assert_frame_equal equivalence check between
the flags path and the metrics= path.


* Tier 3: split validation/analysis/plotting into mixin modules (#363)

Extract the three cohesive, largely-independent concern clusters out of the
2.8k-line evaluator.py into focused modules, mixed back into BenchmarkEvaluator so
the public API is byte-for-byte identical (methods still resolve on the class via
inheritance — no delegators, verbatim bodies, zero behavior change):

- bencheval/_validation.py  ResultsValidationMixin  (verify_data, verify_data_is_dense,
  verify_error, clean_data, fillna_data)
- bencheval/_analysis.py    DatasetAnalysisMixin    (dataset_representativeness,
  dataset_fold_similarity, rank_datasets_by_fold_similarity, the jitter family,
  estimate_folds_for_stable_ordering)
- bencheval/_plotting.py    PlottingMixin           (plot_winrate_matrix,
  plot_critical_diagrams, plot_dataset_metric_distribution)
- bencheval/_common.py      shared column-name constants (breaks the evaluator<->mixin
  import cycle; re-exported from evaluator.py for backward compatibility)

evaluator.py drops from 2832 to 1233 lines and now focuses on construction, the
leaderboard/metric registry, the compute_* metrics, and the greedy MetricSpec.

Mixins (composition-by-inheritance) were chosen over standalone classes + delegation
because ~17 of these methods are public and several are called externally; mixins
preserve that surface exactly with verbatim bodies (lowest risk) while still achieving
the file-level separation. Converting to fully standalone classes is a possible
follow-up if independent reuse is needed.

Tests: the 24 existing behavior/registry tests pass unchanged, plus 7 new smoke tests
that execute every extracted analysis and plotting method (these clusters had no prior
coverage) — so the split is verified end to end, not just at import.


---------
