+++ packages/bencheval/src/bencheval/evaluator.py
@@ -1,27 +1,22 @@
 from __future__ import annotations
 
-import copy
-import os
 from dataclasses import dataclass
-from pathlib import Path
 from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import pandas as pd
-from pandas.api.types import is_numeric_dtype
 from scipy.stats import gmean
 
+from ._analysis import DatasetAnalysisMixin
+from ._common import FRONTIER_ADVANTAGE, IMPROVABILITY, LOSS_RESCALED, RANK
+from ._plotting import PlottingMixin
+from ._validation import ResultsValidationMixin
 from .elo_utils import EloHelper
 from .mean_utils import compute_weighted_mean_by_task
 from .winrate_utils import compute_winrate, compute_winrate_matrix
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable
-
-RANK = "rank"
-IMPROVABILITY = "improvability"
-LOSS_RESCALED = "loss_rescaled"
-FRONTIER_ADVANTAGE = "frontier_advantage"
+    from collections.abc import Callable, Iterable, Sequence
 
 MetricDirection = Literal["min", "max"]
 MetricAlignment = Literal["row", "method"]
@@ -50,8 +45,135 @@ class MetricSpec:
     invalid_subset_policy: InvalidSubsetPolicy = "raise"
 
 
+@dataclass
+class _LeaderboardContext:
+    """Mutable state threaded through the leaderboard metric producers.
+
+    ``results_agg`` may be replaced by a producer (e.g. ``improvability`` pops its own
+    column out of the aggregate and re-emits it with bootstrap CI bars), so the trailing
+    aggregate block is read from the context *after* every producer has run.
+    """
+
+    evaluator: BenchmarkEvaluator
+    results_per_task: pd.DataFrame
+    results_agg: pd.DataFrame
+    baseline_method: str | None
+    elo_kwargs: dict
+    relative_error_kwargs: dict
+
+
+@dataclass(frozen=True)
+class _LeaderboardMetric:
+    """One column-group of a leaderboard, produced from a ``_LeaderboardContext``.
+
+    ``produce`` returns the list of Series/DataFrames to concatenate (in order) for this
+    metric. Adding a leaderboard column is a single registry entry plus a selector key —
+    no new ``leaderboard()`` parameter or ``if`` branch.
+    """
+
+    key: str
+    produce: Callable[[_LeaderboardContext], list]
+    requires_baseline: bool = False
+    always_on: bool = False
+
+
+def _lb_elo(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_elo(results_per_task=ctx.results_per_task, **ctx.elo_kwargs)]
+
+
+def _lb_rank(ctx: _LeaderboardContext) -> list:
+    return [ctx.results_agg[RANK]]
+
+
+def _lb_winrate(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_winrate(results_per_task=ctx.results_per_task).to_frame()]
+
+
+def _lb_improvability(ctx: _LeaderboardContext) -> list:
+    ev = ctx.evaluator
+    tasks = list(ctx.results_per_task[ev.task_col].unique())
+    results_per_task_avg = ctx.results_per_task.groupby(ev.groupby_columns)[IMPROVABILITY].mean().reset_index()
+    improvability_bootstrap = get_bootstrap_result_lst(
+        data=tasks,
+        func_=ev._weighted_groupby_mean,
+        func_kwargs={"data": results_per_task_avg, "agg_column": IMPROVABILITY},
+        num_round=100,
+    )
+    improvability = ctx.results_agg[IMPROVABILITY]
+    ctx.results_agg = ctx.results_agg.drop(columns=[IMPROVABILITY])
+    improvability_quantiles = pd.DataFrame(
+        {
+            f"{IMPROVABILITY}+": improvability_bootstrap.quantile(0.975) - improvability,
+            f"{IMPROVABILITY}-": improvability - improvability_bootstrap.quantile(0.025),
+        }
+    )
+    return [improvability, improvability_quantiles]
+
+
+def _lb_baseline_advantage(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_baseline_advantage(ctx.results_per_task, baseline_method=ctx.baseline_method)]
+
+
+def _lb_frontier_advantage(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_frontier_advantage(results_per_task=ctx.results_per_task)]
+
+
+def _lb_mrr(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_mrr(results_per_task=ctx.results_per_task).to_frame()]
+
+
+def _lb_relative_error(ctx: _LeaderboardContext) -> list:
+    return [
+        ctx.evaluator.compute_relative_error(
+            results_per_task=ctx.results_per_task,
+            baseline_method=ctx.baseline_method,
+            **ctx.relative_error_kwargs,
+        ).to_frame()
+    ]
+
+
+def _lb_skill_score(ctx: _LeaderboardContext) -> list:
+    return [
+        ctx.evaluator.compute_skill_score(results_per_task=ctx.results_per_task, baseline_method=ctx.baseline_method)
+    ]
+
+
+def _lb_rank_counts(ctx: _LeaderboardContext) -> list:
+    return [ctx.evaluator.compute_rank_counts(results_per_task=ctx.results_per_task)]
+
+
+# Ordered registry: the leaderboard emits these column-groups in this order. ``always_on``
+# metrics are always emitted; the rest are selected via ``leaderboard(metrics=...)`` (or the
+# legacy ``include_*`` flags). ``requires_baseline`` metrics are skipped when no baseline is set.
+_LEADERBOARD_METRICS: tuple[_LeaderboardMetric, ...] = (
+    _LeaderboardMetric("elo", _lb_elo),
+    _LeaderboardMetric("rank", _lb_rank, always_on=True),
+    _LeaderboardMetric("winrate", _lb_winrate),
+    _LeaderboardMetric("improvability", _lb_improvability),
+    _LeaderboardMetric("baseline_advantage", _lb_baseline_advantage, requires_baseline=True),
+    _LeaderboardMetric("frontier_advantage", _lb_frontier_advantage),
+    _LeaderboardMetric("mrr", _lb_mrr),
+    _LeaderboardMetric("relative_error", _lb_relative_error, requires_baseline=True),
+    _LeaderboardMetric("skill_score", _lb_skill_score, requires_baseline=True),
+    _LeaderboardMetric("rank_counts", _lb_rank_counts),
+)
+
+# Maps the legacy ``include_*`` flags onto registry keys (back-compat selector layer).
+_LEADERBOARD_FLAG_TO_KEY = {
+    "include_elo": "elo",
+    "include_winrate": "winrate",
+    "include_improvability": "improvability",
+    "include_mrr": "mrr",
+    "include_rank_counts": "rank_counts",
+    "include_relative_error": "relative_error",
+    "include_skill_score": "skill_score",
+    "include_baseline_advantage": "baseline_advantage",
+    "include_frontier_advantage": "frontier_advantage",
+}
+
+
 # TODO: Should "data" be an init arg? Probably not.
-class BenchmarkEvaluator:
+class BenchmarkEvaluator(ResultsValidationMixin, DatasetAnalysisMixin, PlottingMixin):
     def __init__(
         self,
         method_col: str = "method",
@@ -122,7 +244,19 @@ class BenchmarkEvaluator:
         relative_error_kwargs: dict | None = None,
         elo_kwargs: dict | None = None,
         sort_by: str | list[str] | None = "rank",
+        metrics: Sequence[str] | None = None,
     ):
+        """Build a per-method leaderboard.
+
+        Metric columns are produced by the ordered registry ``_LEADERBOARD_METRICS``.
+        Select them either with the legacy ``include_*`` flags or, equivalently, by passing
+        ``metrics`` — an iterable of metric keys (e.g. ``["elo", "winrate", "mrr"]``) which,
+        when given, overrides the flags. ``rank`` is always included. ``baseline_method`` is
+        required for ``baseline_advantage`` / ``relative_error`` / ``skill_score`` (they are
+        silently skipped without it). ``include_error`` / ``include_rescaled_loss`` toggle
+        columns carried over from the aggregate, and ``average_seeds`` controls per-seed
+        handling.
+        """
         if elo_kwargs is None:
             elo_kwargs = {}
         if relative_error_kwargs is None:
@@ -130,74 +264,49 @@ class BenchmarkEvaluator:
         if baseline_method is None:
             baseline_method = elo_kwargs.get("calibration_framework")
 
+        enabled = self._resolve_leaderboard_metrics(
+            metrics,
+            {
+                "include_elo": include_elo,
+                "include_winrate": include_winrate,
+                "include_improvability": include_improvability,
+                "include_mrr": include_mrr,
+                "include_rank_counts": include_rank_counts,
+                "include_relative_error": include_relative_error,
+                "include_skill_score": include_skill_score,
+                "include_baseline_advantage": include_baseline_advantage,
+                "include_frontier_advantage": include_frontier_advantage,
+            },
+        )
+
         self.verify_data(data=data)
 
-        if average_seeds:
-            # average each method's task error across the seeds
-            # Calculate all metrics on the averaged error for the task.
-            results_per_task = self.compute_results_per_task(data=data)
-        else:
-            # Keep each method's task error for each seed, don't average the error.
-            # Calculate all metrics on each seed, then average across seeds to get the metric value for the task.
-            results_per_task = self.compute_results_per_task(data=data, include_seed_col=True)
+        # average_seeds=True averages each method's per-task error across seeds first;
+        # otherwise metrics are computed per seed and averaged afterwards.
+        results_per_task = self.compute_results_per_task(data=data, include_seed_col=not average_seeds)
 
-        results_agg = self.aggregate(results_by_dataset=results_per_task)
+        ctx = _LeaderboardContext(
+            evaluator=self,
+            results_per_task=results_per_task,
+            results_agg=self.aggregate(results_by_dataset=results_per_task),
+            baseline_method=baseline_method,
+            elo_kwargs=elo_kwargs,
+            relative_error_kwargs=relative_error_kwargs,
+        )
+
         results_lst = []
+        for metric in _LEADERBOARD_METRICS:
+            if not metric.always_on and metric.key not in enabled:
+                continue
+            if metric.requires_baseline and baseline_method is None:
+                continue
+            results_lst.extend(metric.produce(ctx))
 
-        if include_elo:
-            results_lst.append(self.compute_elo(results_per_task=results_per_task, **elo_kwargs))
-        results_lst.append(results_agg[RANK])
-        if include_winrate:
-            results_lst.append(self.compute_winrate(results_per_task=results_per_task).to_frame())
-        if include_improvability:
-            tasks = list(results_per_task[self.task_col].unique())
-            results_per_task_avg = results_per_task.groupby(self.groupby_columns)[IMPROVABILITY].mean().reset_index()
-            improvability_bootstrap = get_bootstrap_result_lst(
-                data=tasks,
-                func_=self._weighted_groupby_mean,
-                func_kwargs={"data": results_per_task_avg, "agg_column": IMPROVABILITY},
-                num_round=100,
-            )
-            improvability = results_agg[IMPROVABILITY]
-            results_agg = results_agg.drop(columns=[IMPROVABILITY])
-            improvability_quantiles = pd.DataFrame(
-                {
-                    f"{IMPROVABILITY}+": improvability_bootstrap.quantile(0.975) - improvability,
-                    f"{IMPROVABILITY}-": improvability - improvability_bootstrap.quantile(0.025),
-                }
-            )
-
-            results_lst += [improvability, improvability_quantiles]
-        if include_baseline_advantage and baseline_method is not None:
-            results_lst.append(
-                self.compute_baseline_advantage(
-                    results_per_task,
-                    baseline_method=baseline_method,
-                )
-            )
-        if include_frontier_advantage:
-            results_lst.append(self.compute_frontier_advantage(results_per_task=results_per_task))
-        if include_mrr:
-            results_lst.append(self.compute_mrr(results_per_task=results_per_task).to_frame())
-        if baseline_method is not None:
-            if include_relative_error:
-                results_lst.append(
-                    self.compute_relative_error(
-                        results_per_task=results_per_task,
-                        baseline_method=baseline_method,
-                        **relative_error_kwargs,
-                    ).to_frame(),
-                )
-            if include_skill_score:
-                results_lst.append(
-                    self.compute_skill_score(results_per_task=results_per_task, baseline_method=baseline_method),
-                )
-
-        if include_rank_counts:
-            results_lst.append(self.compute_rank_counts(results_per_task=results_per_task))
-
-        cols_to_use = [c for c in results_agg.columns if c != RANK]
-        results_lst.append(results_agg[cols_to_use])
+        # Trailing block: every aggregated column except rank (emitted above). A producer may
+        # have popped its own column out of ctx.results_agg (e.g. improvability), so read it
+        # here, after the loop.
+        cols_to_use = [c for c in ctx.results_agg.columns if c != RANK]
+        results_lst.append(ctx.results_agg[cols_to_use])
 
         results = pd.concat(results_lst, axis=1)
 
@@ -207,247 +316,35 @@ class BenchmarkEvaluator:
             results = results.drop(columns=[self.error_col])
         if not include_rescaled_loss:
             results = results.drop(columns=[LOSS_RESCALED])
-        if not include_improvability:
+        if "improvability" not in enabled:
             results = results.drop(columns=[IMPROVABILITY])
         results.index.name = self.method_col
 
         return results
 
-    def verify_data(self, data: pd.DataFrame):
-        assert isinstance(data, pd.DataFrame)
-        data_columns = list(data.columns)
-        data_columns_set = set(data_columns)
-        assert len(data_columns) == len(data_columns_set)
+    @staticmethod
+    def _resolve_leaderboard_metrics(metrics: Sequence[str] | None, include_flags: dict[str, bool]) -> set[str]:
+        """Resolve the set of enabled metric keys from ``metrics`` or the ``include_*`` flags.
 
-        missing_columns = []
-        present_columns = []
-        for c in self.columns_to_agg:
-            if c not in data_columns_set:
-                missing_columns.append(c)
-            else:
-                present_columns.append(c)
-        for c in self.groupby_columns:
-            if c not in data_columns_set:
-                missing_columns.append(c)
-            else:
-                present_columns.append(c)
-        if self.seed_column is not None:
-            if self.seed_column not in data_columns_set:
-                missing_columns.append(self.seed_column)
-            else:
-                present_columns.append(self.seed_column)
-
-        required_columns = self.groupby_columns + self.columns_to_agg
-        if self.seed_column is not None:
-            required_columns.append(self.seed_column)
-        unused_columns = [d for d in data_columns if d not in required_columns]
-
-        if missing_columns:
-            index_names = data.index.names
-            missing_in_index = []
-            for index_name in index_names:
-                if index_name in missing_columns:
-                    missing_in_index.append(index_name)
-            if missing_in_index:
-                msg_extra = (
-                    "Columns exist in the index that are required to be columns! "
-                    "\n\tEnsure you reset your index to make these columns available: `data = data.reset_index()`\n"
+        ``metrics`` (when not ``None``) takes precedence and is validated against the registry;
+        otherwise the legacy boolean flags select the keys. ``rank`` is always emitted and is
+        not part of this set.
+        """
+        all_keys = {m.key for m in _LEADERBOARD_METRICS}
+        if metrics is not None:
+            requested = set(metrics)
+            unknown = requested - all_keys
+            if unknown:
+                raise ValueError(
+                    f"Unknown leaderboard metric(s): {sorted(unknown)}. Available: {sorted(all_keys)}.",
                 )
-            else:
-                msg_extra = ""
-            raise ValueError(
-                f"{msg_extra}"
-                f"Missing required columns:"
-                f"\n\tMissing columns ({len(missing_columns)}): {missing_columns}"
-                f"\n\tExisting columns ({len(present_columns)}): {present_columns}"
-                f"\n\tUnused columns ({len(unused_columns)}): {unused_columns}"
-                f"\n\tIndex names ({len(index_names)}): {index_names}",
-            )
-
-        for c in self.groupby_columns:
-            assert data[c].isnull().sum() == 0, f"groupby column {c!r} contains NaN!"
-        for c in self.columns_to_agg:
-            assert is_numeric_dtype(data[c]), "aggregation columns must be numeric!"
-        for c in self.columns_to_agg:
-            if data[c].isnull().sum() != 0:
-                invalid_samples = data[data[c].isnull()]
-
-                raise AssertionError(
-                    f"Column {c} should not contain null values. "
-                    f"Found {data[c].isnull().sum()}/{len(data)} null values! "
-                    f"Invalid samples:\n{invalid_samples.head(100).to_markdown()}",
-                )
-
-        # TODO: Check no duplicates
-        len_data = len(data)
-        unique_val_columns = [self.task_col, self.method_col]
-        if self.seed_column is not None:
-            unique_val_columns.append(self.seed_column)
-        len_data_dedupe = len(data.drop_duplicates(unique_val_columns))
-        assert len_data == len_data_dedupe
-
-        self.verify_data_is_dense(data=data)
-        self.verify_error(data=data)
-
-    def verify_data_is_dense(self, data: pd.DataFrame):
-        methods = list(data[self.method_col].unique())
-        num_methods = len(methods)
-        # FIXME: seed_column
-        datasets = list(data[self.task_col].unique())
-        num_datasets = len(datasets)
-
-        task_cols = self.get_task_groupby_cols(include_seed_col=True)
-        unique_tasks = data[task_cols].drop_duplicates().reset_index(drop=True)
-
-        unique_seeds_per_dataset = unique_tasks[self.task_col].value_counts()
-        num_tasks = unique_seeds_per_dataset.sum()
-        valid_tasks_per_method = data[self.method_col].value_counts()
-        valid_methods_per_dataset = data[self.task_col].value_counts()
-        valid_methods_per_task = data[task_cols].value_counts()
-        invalid_tasks_per_method = (-valid_tasks_per_method + num_tasks).sort_values(ascending=False)
-        invalid_methods_per_dataset = (
-            -valid_methods_per_dataset + valid_methods_per_dataset.index.map(unique_seeds_per_dataset) * num_methods
-        ).sort_values(ascending=False)
-        invalid_methods_per_task = (-valid_methods_per_task + num_methods).sort_values(ascending=False)
-
-        if (invalid_tasks_per_method != 0).any():
-            invalid_tasks_per_method_filtered = invalid_tasks_per_method[invalid_tasks_per_method != 0]
-            invalid_methods_per_dataset_filtered = invalid_methods_per_dataset[invalid_methods_per_dataset != 0]
-            invalid_methods_per_task_filtered = invalid_methods_per_task[invalid_methods_per_task != 0]
-            num_invalid_results = invalid_tasks_per_method.sum()
-            # num_invalid_tasks = invalid_methods_per_task_filtered.sum()
-
-            df_experiments_dense = unique_tasks.merge(
-                pd.Series(data=methods, name=self.method_col),
-                how="cross",
-            )
-            experiment_cols = [*task_cols, self.method_col]
-            overlap = pd.merge(
-                df_experiments_dense, data[experiment_cols], on=experiment_cols, how="left", indicator="exist"
-            )
-            df_missing_experiments = (
-                overlap[overlap["exist"] == "left_only"][experiment_cols]
-                .sort_values(by=experiment_cols)
-                .reset_index(drop=True)
-            )
-
-            with pd.option_context("display.max_rows", None, "display.max_columns", None, "display.width", 1000):
-                if len(df_missing_experiments) <= 500:
-                    print(f"\nFailed Experiments ({len(df_missing_experiments)}):")
-                    print(df_missing_experiments)
-                print("\nMethods sorted by failure count:")
-                print(invalid_tasks_per_method_filtered)
-                print("\nDatasets sorted by failure count:")
-                print(invalid_methods_per_dataset_filtered)
-            # missing results
-            raise AssertionError(
-                f"Missing results for some methods. Ensure that all methods have results for all tasks.\n"
-                f"If failures exist, fill missing values before passing into this method.\n"
-                f"{len(invalid_tasks_per_method_filtered)}/{num_methods} methods with missing tasks. {num_invalid_results} missing results.\n"
-                f"{len(invalid_methods_per_dataset_filtered)}/{num_datasets} datasets with missing methods.\n"
-                f"{len(invalid_methods_per_task_filtered)}/{num_tasks} tasks with missing methods.\n"
-                f"Methods sorted by failure count:\n"
-                f"{invalid_tasks_per_method_filtered}",
-            )
-
-    def verify_error(self, data: pd.DataFrame):
-        min_error = data[self.error_col].min()
-        if min_error < 0:
-            data_invalid = data[data[self.error_col] < 0]
-            num_invalid = len(data_invalid)
-            raise ValueError(
-                f"Found {num_invalid} rows where {self.error_col} is less than 0! Error can never be less than 0. "
-                f"Ensure your error is computed correctly."
-                f"\nMinimum value found: {min_error}"
-                f"\nSometimes floating point precision can result in a tiny negative value. "
-                f"You can fix this by doing: data['{self.error_col}'] = data['{self.error_col}'].clip(lower=0)",
-            )
+            return requested
+        return {_LEADERBOARD_FLAG_TO_KEY[flag] for flag, on in include_flags.items() if on}
 
     # TODO: Consider moving this to a different class or finding a better separation.
     #  The eval code becomes a lot more complicated if we need to account for improperly formatted / invalid data.
-    def clean_data(
-        self,
-        data: pd.DataFrame,
-    ) -> pd.DataFrame:
-        data = copy.deepcopy(data)
-        min_error = data[self.error_col].min()
-        if min_error < 0:
-            if min_error >= self.negative_error_threshold:
-                data[self.error_col] = data[self.error_col].clip(0)
-            else:
-                self.verify_error(data=data)
-        return data
 
     # FIXME: Cleanup
-    def fillna_data(
-        self,
-        data: pd.DataFrame,
-        df_fillna: pd.DataFrame | None = None,
-        fillna_method: str = "worst",
-    ) -> pd.DataFrame:
-        """Fills missing (task, seed, method) rows in data with the (task, seed) row in df_fillna.
-
-        Parameters
-        ----------
-        data : pd.DataFrame
-            The data to fill.
-        df_fillna : pd.DataFrame | None, default None
-            If specified, will fill methods with missing results in `data` with the results in `df_fillna`.
-            If specified, `fillna_method` is ignored.
-        fillna_method : str, default "worst"
-            Either "worst" or the name of a method in self.method_col.
-            If "worst", will fill with the result of the method with the worst error on a given task.
-            Ignored if `df_fillna` is specified.
-
-        Returns:
-        -------
-        pd.DataFrame
-            The filled data.
-
-        """
-        task_columns = [self.task_col, self.seed_column] if self.seed_column else [self.task_col]
-
-        unique_methods = list(data[self.method_col].unique())
-
-        if df_fillna is None:
-            if fillna_method == "worst":
-                assert df_fillna is None, "df_fillna must be None if fillna_method='worst'"
-                idx_worst = data.groupby(task_columns)[self.error_col].idxmax()
-                df_fillna = data.loc[idx_worst]
-            elif isinstance(fillna_method, str) and fillna_method in data[self.method_col].unique():
-                df_fillna = data.loc[data[self.method_col] == fillna_method]
-            else:
-                raise AssertionError(
-                    f"df_fillna is None and fillna_method {fillna_method!r} is not present in data."
-                    f"\n\tValid methods: {list(data[self.method_col].unique())}",
-                )
-        if self.method_col in df_fillna.columns:
-            df_fillna = df_fillna.drop(columns=[self.method_col])
-
-        data = data.set_index([*task_columns, self.method_col], drop=True)
-
-        df_filled = df_fillna[task_columns].merge(
-            pd.Series(data=unique_methods, name=self.method_col),
-            how="cross",
-        )
-        df_filled = df_filled.set_index(keys=list(df_filled.columns))
-
-        # missing results
-        nan_vals = df_filled.index.difference(data.index)
-
-        # fill valid values
-        fill_cols = list(data.columns)
-        df_filled[fill_cols] = np.nan
-        df_filled[fill_cols] = df_filled[fill_cols].astype(data.dtypes)
-        df_filled.loc[data.index] = data
-
-        df_fillna = df_fillna.set_index(task_columns, drop=True)
-        a = df_fillna.loc[nan_vals.droplevel(level=self.method_col)]
-        a.index = nan_vals
-        df_filled.loc[nan_vals] = a
-        data = df_filled
-
-        return data.reset_index(drop=False)
 
     def get_task_groupby_cols(self, include_seed_col: bool = False):
         task_groupby_cols = self.task_groupby_columns
@@ -847,116 +744,6 @@ class BenchmarkEvaluator:
             seed_col=seed_col,
         )
 
-    @staticmethod
-    def plot_winrate_matrix(
-        winrate_matrix: pd.DataFrame,
-        save_path: str | None,
-        title: str | None = None,
-    ):
-        import matplotlib.pyplot as plt
-
-        z = winrate_matrix.copy()
-        z = (z * 100).round().astype("Int64")
-
-        n_rows, n_cols = z.shape
-
-        # Scale figure size with matrix size, while keeping a sensible minimum
-        fig_w = max(11.1, 0.55 * n_cols + 3)
-        fig_h = max(9.0, 0.45 * n_rows + 2)
-
-        fig, ax = plt.subplots(figsize=(fig_w, fig_h), constrained_layout=True)
-
-        # PRGn-like colormap
-        im = ax.imshow(z.astype(float).to_numpy(), cmap="PRGn", vmin=0, vmax=100)
-
-        # Ticks / labels
-        ax.set_xticks(range(n_cols))
-        ax.set_yticks(range(n_rows))
-        ax.set_xticklabels(
-            z.columns,
-            fontsize=16,
-            rotation=330,
-            ha="right",
-            va="bottom",
-            rotation_mode="anchor",
-        )
-        ax.set_yticklabels(z.index, fontsize=16)
-        ax.tick_params(axis="x", pad=-2)
-        ax.tick_params(axis="x", which="both", length=0)
-        ax.tick_params(axis="y", which="both", length=0)
-
-        ax.xaxis.tick_top()
-        ax.xaxis.set_label_position("top")
-        ax.set_xlabel("Model B: Loser", fontsize=18)
-        ax.set_ylabel("Model A: Winner", fontsize=18)
-
-        cmap = plt.get_cmap("PRGn")
-        norm = plt.Normalize(vmin=0, vmax=100)
-
-        values = z.to_numpy()
-        for i in range(n_rows):
-            for j in range(n_cols):
-                val = values[i, j]
-                if pd.isna(val):
-                    text = ""
-                    text_color = "black"
-                else:
-                    text = f"{int(val)}"
-                    r, g, b, _ = cmap(norm(float(val)))
-                    luminance = 0.2126 * r + 0.7152 * g + 0.0722 * b
-                    text_color = "black" if luminance > 0.5 else "white"
-
-                ax.text(
-                    j,
-                    i,
-                    text,
-                    ha="center",
-                    va="center",
-                    fontsize=16,
-                    color=text_color,
-                )
-
-        # Colorbar
-        from matplotlib.ticker import FuncFormatter
-
-        cbar = fig.colorbar(im, ax=ax)
-        cbar.ax.set_title("Win Rate", fontsize=18, pad=12, loc="left")
-        cbar.ax.tick_params(labelsize=16)
-        cbar.ax.yaxis.set_major_formatter(FuncFormatter(lambda x, pos: f"{int(x)}%"))
-
-        # Clean look
-        for spine in ax.spines.values():
-            spine.set_visible(False)
-
-        ax.set_facecolor("white")
-        fig.patch.set_facecolor("white")
-
-        # Match imshow orientation expectations
-        ax.set_ylim(n_rows - 0.5, -0.5)
-
-        if title is not None:
-            # Center horizontally on the *grid* (i.e. the imshow axes), the
-            # same way ``ax.set_xlabel("Model B: Loser", ...)`` is centered.
-            # ``ax.get_position()`` returns the axes bounding box in figure
-            # coordinates, so the midpoint of its x-range is the matrix
-            # grid's horizontal center — independent of the colorbar that
-            # sits to the right.
-            fig.canvas.draw()
-            ax_bbox = ax.get_position()
-            x_axes_center = (ax_bbox.x0 + ax_bbox.x1) / 2
-            fig.suptitle(
-                title,
-                fontsize=20,
-                fontweight="bold",
-                x=x_axes_center,
-            )
-
-        if save_path is not None:
-            if os.path.dirname(save_path):
-                os.makedirs(os.path.dirname(save_path), exist_ok=True)
-            fig.savefig(save_path, bbox_inches="tight")
-        plt.close(fig)
-
     def compare_rank_per(
         self,
         df: pd.DataFrame,
@@ -1124,38 +911,6 @@ class BenchmarkEvaluator:
         )
         results_rank.name = RANK
         return results_rank
-
-    # TODO: Should plotting live in a separate class?
-    def plot_critical_diagrams(
-        self, results_per_task, save_path: str | None = None, show: bool = False, reverse: bool = False
-    ):
-        import matplotlib.pyplot as plt
-
-        with plt.rc_context({"text.usetex": False}):
-            from autorank import autorank
-            from autorank._util import cd_diagram
-
-            plt.rcParams.update({"font.size": 12})
-
-            fig = plt.figure(figsize=(10, 4))
-            ax = fig.add_subplot(1, 1, 1)
-
-            data = results_per_task.pivot_table(index=self.task_col, columns=self.method_col, values="rank")
-            result = autorank(data, alpha=0.05, verbose=False, order="ascending", force_mode="nonparametric")
-
-            try:
-                _ = cd_diagram(result, reverse=reverse, ax=ax, width=6)
-            except KeyError:
-                print("Not enough methods to generate cd_diagram, skipping...")
-                return
-
-            # plt.tight_layout()  # cuts off text
-            if save_path is not None:
-                parent_dir = str(Path(save_path).parent)
-                os.makedirs(parent_dir, exist_ok=True)
-                plt.savefig(save_path, bbox_inches="tight", pad_inches=0.1)
-            if show:
-                plt.show()
 
     # TODO: Make faster, can be 100x faster if vectorized properly.
     def _weighted_groupby_mean(self, tasks: list[str], data: pd.DataFrame, agg_column: str) -> pd.Series:
@@ -1457,1227 +1212,6 @@ class BenchmarkEvaluator:
             return False
         # "skip"
         return False
-
-    def plot_dataset_metric_distribution(
-        self,
-        results_per_task: pd.DataFrame,
-        dataset: str,
-        metric_col: str | None = None,
-        *,
-        sort_by: str = "median",  # {"median", "mean"}
-        ascending: bool = True,  # lower is better for error-like metrics
-        kind: str = "violin",  # {"violin", "box", "bar"}
-        show_points: str | bool = "outliers",  # plotly: True, False, "all", "outliers"
-        log_y: bool = False,
-        max_methods: int | None = 50,  # cap for readability
-        save_path: str | None = None,
-        title: str | None = None,
-    ):
-        """Plot a single dataset's metric distribution across methods.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-            Output of `self.compute_results_per_task(data=data)` (possibly with seeds).
-            Must contain at least: [self.task_col, self.method_col, metric_col].
-        dataset : str
-            Dataset/task name to filter on (value from `self.task_col`).
-        metric_col : str | None, default None
-            Column to plot. If None, defaults to `self.error_col`.
-            Common choices: self.error_col, "rank", "improvability", "loss_rescaled".
-        sort_by : {"median","mean"}, default "median"
-            How to sort methods on the x-axis.
-        ascending : bool, default True
-            Whether to sort ascending (typical for error-like metrics).
-        kind : {"violin","box","bar"}, default "violin"
-            Plot type. If the dataset has only one value per method, it will fall back to "bar".
-        show_points : str | bool, default "outliers"
-            Whether to show individual points for violin/box.
-        log_y : bool, default False
-            Whether to log-scale the y-axis.
-        max_methods : int | None, default 50
-            If set, keep only the top-N methods by the chosen sorter (post-sort).
-        save_path : str | None, default None
-            If provided, writes the figure to this path (format inferred by extension).
-        title : str | None, default None
-            Optional plot title.
-
-        Returns:
-        -------
-        fig : plotly.graph_objects.Figure
-        df_plot : pd.DataFrame
-            Filtered dataframe used for plotting (one row per observation).
-        """
-        import plotly.express as px
-
-        if metric_col is None:
-            metric_col = self.error_col
-
-        required = {self.task_col, self.method_col, metric_col}
-        missing = [c for c in required if c not in results_per_task.columns]
-        if missing:
-            raise ValueError(
-                f"results_per_task is missing required columns: {missing}\n"
-                f"Available columns: {list(results_per_task.columns)}",
-            )
-
-        df = results_per_task.loc[results_per_task[self.task_col] == dataset].copy()
-        if df.empty:
-            raise ValueError(
-                f"No rows found for {self.task_col} == {dataset!r}.\n"
-                f"Available datasets: {sorted(results_per_task[self.task_col].unique())[:50]}",
-            )
-
-        # Drop NaNs in the plotted metric (shouldn’t normally exist, but be safe)
-        df = df.dropna(subset=[metric_col])
-
-        # Determine whether we actually have a distribution per method (e.g. seeds)
-        # If each method has a single value, violin/box isn't very informative.
-        per_method_counts = df.groupby(self.method_col)[metric_col].size()
-        has_distribution = bool((per_method_counts > 1).any())
-
-        # Sort methods by mean/median for consistent x-axis ordering
-        if sort_by not in {"median", "mean"}:
-            raise ValueError(f"Invalid sort_by={sort_by!r}. Expected 'median' or 'mean'.")
-
-        if sort_by == "median":
-            sorter = df.groupby(self.method_col)[metric_col].median()
-        else:
-            sorter = df.groupby(self.method_col)[metric_col].mean()
-
-        sorter = sorter.sort_values(ascending=ascending)
-
-        if max_methods is not None and len(sorter) > max_methods:
-            sorter = sorter.iloc[:max_methods]
-            df = df[df[self.method_col].isin(sorter.index)].copy()
-
-        method_order = list(sorter.index)
-
-        # Pick plot type (force bar if no distribution)
-        plot_kind = kind
-        if plot_kind in {"violin", "box"} and not has_distribution:
-            plot_kind = "bar"
-
-        if title is None:
-            title = f"{dataset}: {metric_col} across methods"
-
-        if plot_kind == "violin":
-            fig = px.violin(
-                df,
-                x=self.method_col,
-                y=metric_col,
-                category_orders={self.method_col: method_order},
-                box=True,
-                points=show_points,
-                title=title,
-            )
-        elif plot_kind == "box":
-            fig = px.box(
-                df,
-                x=self.method_col,
-                y=metric_col,
-                category_orders={self.method_col: method_order},
-                points=show_points,
-                title=title,
-            )
-        elif plot_kind == "bar":
-            # One value per method (or user forced bar): use median/mean as height
-            agg = sorter.rename(metric_col).reset_index()
-            fig = px.bar(
-                agg,
-                x=self.method_col,
-                y=metric_col,
-                category_orders={self.method_col: method_order},
-                title=title,
-            )
-        else:
-            raise ValueError(f"Invalid kind={kind!r}. Expected 'violin', 'box', or 'bar'.")
-
-        fig.update_layout(
-            xaxis_title="Method",
-            yaxis_title=metric_col,
-            xaxis_tickangle=45,
-            margin=dict(l=10, r=10, t=40, b=10),
-            plot_bgcolor="white",
-        )
-        fig.update_xaxes(showgrid=False)
-        fig.update_yaxes(showgrid=True)
-
-        if log_y:
-            fig.update_yaxes(type="log")
-
-        if save_path is not None:
-            if os.path.dirname(save_path):
-                os.makedirs(os.path.dirname(save_path), exist_ok=True)
-            fig.write_image(save_path)
-
-        return fig, df
-
-    def dataset_representativeness(
-        self,
-        results_per_task: pd.DataFrame,
-        value_col: str = RANK,
-        *,
-        similarity: str = "spearman",  # {"spearman", "pearson"}
-        population_mode: str = "loo_mean",  # {"loo_mean", "global_mean"}
-        score_mode: str = "to_population",  # {"to_population", "mean_pairwise"}
-        min_methods: int | None = None,
-    ) -> dict:
-        """Quantify how similar each dataset is to the others based on model performance,
-        then identify the most- and least-representative datasets.
-
-        Intuition
-        ---------
-        Each dataset defines a vector over methods (e.g., ranks or errors).
-        Two datasets are "similar" if these vectors are correlated across methods.
-        Representativeness is measured by how well a dataset aligns with the population.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-            Output of `self.compute_results_per_task(data=data)` (may include seed col).
-        value_col : str, default "rank"
-            Column used to compare method performance. Common:
-            - RANK (lower is better)
-            - self.error_col (lower is better)
-            - IMPROVABILITY (higher is better) [still fine; correlation handles direction]
-            - LOSS_RESCALED (lower is better)
-        similarity : {"spearman","pearson"}, default "spearman"
-            Similarity metric for comparing datasets via correlation across methods.
-            Spearman is recommended (rank-based; robust to scale).
-        population_mode : {"loo_mean","global_mean"}, default "loo_mean"
-            How to define the population reference vector:
-            - "loo_mean": reference for dataset d is mean vector over all datasets except d
-            - "global_mean": reference is mean vector over all datasets (includes d)
-        score_mode : {"to_population","mean_pairwise"}, default "to_population"
-            How to assign a single representativeness score per dataset:
-            - "to_population": correlation(dataset_vector, population_reference_vector)
-            - "mean_pairwise": average correlation to all other datasets
-        min_methods : int | None
-            If set, require at least this many methods to compute correlations.
-
-        Returns:
-        -------
-        out : dict with keys
-            - "representativeness": pd.DataFrame indexed by dataset with columns:
-                ["score", "num_methods", "num_obs"]
-            - "most_representative": str
-            - "least_representative": str
-            - "dataset_similarity_matrix": pd.DataFrame (dataset x dataset)
-            - "method_matrix": pd.DataFrame (dataset x method), values = value_col (after seed-avg)
-        """
-        if similarity not in {"spearman", "pearson"}:
-            raise ValueError(f"similarity must be 'spearman' or 'pearson', got {similarity!r}")
-        if population_mode not in {"loo_mean", "global_mean"}:
-            raise ValueError(f"population_mode must be 'loo_mean' or 'global_mean', got {population_mode!r}")
-        if score_mode not in {"to_population", "mean_pairwise"}:
-            raise ValueError(f"score_mode must be 'to_population' or 'mean_pairwise', got {score_mode!r}")
-
-        required = {self.task_col, self.method_col, value_col}
-        missing = [c for c in required if c not in results_per_task.columns]
-        if missing:
-            raise ValueError(
-                f"results_per_task is missing required columns: {missing}\n"
-                f"Available columns: {list(results_per_task.columns)}",
-            )
-
-        df = results_per_task.copy()
-
-        # If seed column exists, average within (task, method) so each dataset has one value per method.
-        group_cols = [self.task_col, self.method_col]
-        if self.seed_column is not None and self.seed_column in df.columns:
-            df = df.groupby(group_cols, as_index=False)[value_col].mean()
-
-        # Pivot to dataset x method matrix
-        M = df.pivot(index=self.task_col, columns=self.method_col, values=value_col)
-
-        # Optional sanity: require sufficient overlap
-        num_methods_per_dataset = M.notna().sum(axis=1)
-        if min_methods is not None:
-            keep = num_methods_per_dataset >= min_methods
-            M = M.loc[keep]
-            num_methods_per_dataset = num_methods_per_dataset.loc[keep]
-            if M.shape[0] == 0:
-                raise ValueError(f"No datasets have >= {min_methods} methods after filtering.")
-
-        # If you're using BenchmarkEvaluator's verify_data_is_dense, M should be fully dense.
-        # But we handle missingness by using pairwise corr + aligning vectors where needed.
-
-        # 1) Dataset↔dataset similarity matrix (correlation across methods)
-        # corr over rows => easiest via transpose (methods as rows, datasets as cols)
-        dataset_sim = M.T.corr(method=similarity)
-
-        # 2) Representativeness score per dataset
-        if score_mode == "mean_pairwise":
-            # Mean similarity to all other datasets (exclude self)
-            score = (
-                (dataset_sim.sum(axis=1) - 1.0) / (dataset_sim.shape[0] - 1)
-                if dataset_sim.shape[0] > 1
-                else dataset_sim.iloc[:, 0]
-            )
-        # Similarity to a "population reference vector"
-        # Reference vector is mean performance across datasets (optionally leave-one-out).
-        elif population_mode == "global_mean":
-            ref = M.mean(axis=0)
-            score = M.apply(lambda row: row.corr(ref, method=similarity), axis=1)
-        else:
-            # Leave-one-out mean reference: ref_d = mean of all datasets except d
-            # Compute efficiently: total_sum - row, then divide by (n-1)
-            n = M.shape[0]
-            if n <= 1:
-                score = pd.Series(index=M.index, data=np.nan, dtype=float)
-            else:
-                total = M.sum(axis=0)
-
-                def loo_corr(row: pd.Series) -> float:
-                    ref_d = (total - row) / (n - 1)
-                    return row.corr(ref_d, method=similarity)
-
-                score = M.apply(loo_corr, axis=1)
-
-        rep = pd.DataFrame(
-            {
-                "score": score,
-                "num_methods": num_methods_per_dataset,
-                "num_obs": df.groupby(self.task_col).size().reindex(M.index).astype(int),
-            }
-        ).sort_values("score", ascending=False)
-
-        most_rep = rep.index[0]
-        least_rep = rep.index[-1]
-
-        return {
-            "representativeness": rep,
-            "most_representative": most_rep,
-            "least_representative": least_rep,
-            "dataset_similarity_matrix": dataset_sim.loc[M.index, M.index],
-            "method_matrix": M,
-        }
-
-    def dataset_fold_similarity(
-        self,
-        results_per_task: pd.DataFrame,
-        dataset: str,
-        value_col: str = RANK,
-        *,
-        similarity: str = "spearman",  # {"spearman", "pearson"}
-        agg_across_methods: str = "mean",  # {"mean", "median"}
-        min_methods: int | None = None,
-        return_pairwise: bool = True,
-    ) -> dict:
-        """Quantify how similar a dataset's folds/seeds are, based on how methods perform.
-
-        Assumes `results_per_task` was produced with `include_seed_col=True`, so that
-        each row corresponds to (task, seed, method) with a metric like rank/error.
-
-        Approach
-        --------
-        For the chosen dataset, build a matrix:
-            rows = fold/seed
-            cols = method
-            values = value_col (e.g., rank or metric_error)
-
-        Then compute fold↔fold similarity as correlation across methods.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-            Output of `self.compute_results_per_task(..., include_seed_col=True)`.
-            Must include `self.seed_column` and `value_col`.
-        dataset : str
-            Task/dataset name (value from `self.task_col`) to analyze.
-        value_col : str, default RANK
-            Performance column to compare across folds. Good defaults:
-            - RANK (ranking alignment)
-            - self.error_col (raw metric error)
-            - IMPROVABILITY, LOSS_RESCALED, etc.
-        similarity : {"spearman","pearson"}, default "spearman"
-            Correlation type for fold similarity.
-        agg_across_methods : {"mean","median"}, default "mean"
-            How to aggregate a per-fold similarity score from its similarities to other folds.
-        min_methods : int | None
-            If set, require each fold have at least this many methods present
-            (after pivot) to be included.
-        return_pairwise : bool, default True
-            If True, also return a tidy dataframe of fold-pair similarities.
-
-        Returns:
-        -------
-        out : dict
-            - "fold_similarity_matrix": pd.DataFrame (fold x fold)
-            - "fold_scores": pd.DataFrame indexed by fold with columns:
-                ["score", "num_methods"]
-            - "most_similar_pair": tuple | None  (fold_i, fold_j, sim)
-            - "least_similar_pair": tuple | None (fold_i, fold_j, sim)
-            - "fold_method_matrix": pd.DataFrame (fold x method)
-            - "pairwise": pd.DataFrame (optional) columns [seed_i, seed_j, similarity]
-        """
-        if self.seed_column is None:
-            raise ValueError("BenchmarkEvaluator.seed_column is None, but fold similarity requires a seed/fold column.")
-        if self.seed_column not in results_per_task.columns:
-            raise ValueError(
-                f"results_per_task must include seed column {self.seed_column!r}. "
-                "Ensure you called compute_results_per_task(..., include_seed_col=True).",
-            )
-        if similarity not in {"spearman", "pearson"}:
-            raise ValueError(f"similarity must be 'spearman' or 'pearson', got {similarity!r}")
-        if agg_across_methods not in {"mean", "median"}:
-            raise ValueError(f"agg_across_methods must be 'mean' or 'median', got {agg_across_methods!r}")
-
-        required = {self.task_col, self.method_col, self.seed_column, value_col}
-        missing = [c for c in required if c not in results_per_task.columns]
-        if missing:
-            raise ValueError(
-                f"results_per_task is missing required columns: {missing}\n"
-                f"Available columns: {list(results_per_task.columns)}",
-            )
-
-        df = results_per_task.loc[results_per_task[self.task_col] == dataset].copy()
-        if df.empty:
-            raise ValueError(
-                f"No rows found for {self.task_col} == {dataset!r}.\n"
-                f"Available datasets: {sorted(results_per_task[self.task_col].unique())[:50]}",
-            )
-
-        # If duplicates exist for (task, seed, method), average them.
-        df = df.groupby([self.task_col, self.seed_column, self.method_col], as_index=False)[value_col].mean()
-
-        # Pivot -> folds x methods
-        M = df.pivot(index=self.seed_column, columns=self.method_col, values=value_col)
-
-        # Optionally filter folds with insufficient method coverage
-        num_methods_per_fold = M.notna().sum(axis=1)
-        if min_methods is not None:
-            keep = num_methods_per_fold >= min_methods
-            M = M.loc[keep]
-            num_methods_per_fold = num_methods_per_fold.loc[keep]
-            if M.shape[0] == 0:
-                raise ValueError(f"No folds have >= {min_methods} methods after filtering.")
-
-        # Correlation across methods between folds
-        # corr over rows => transpose to make folds columns
-        fold_sim = M.T.corr(method=similarity)
-
-        # Per-fold "representativeness among folds": average similarity to other folds
-        if fold_sim.shape[0] <= 1:
-            scores = pd.Series(index=fold_sim.index, data=np.nan, dtype=float)
-        else:
-            off_diag = fold_sim.copy()
-            np.fill_diagonal(off_diag.values, np.nan)
-            if agg_across_methods == "mean":
-                scores = off_diag.mean(axis=1, skipna=True)
-            else:
-                scores = off_diag.median(axis=1, skipna=True)
-
-        fold_scores = pd.DataFrame(
-            {
-                "score": scores,
-                "num_methods": num_methods_per_fold.reindex(fold_sim.index).astype(int),
-            }
-        ).sort_values("score", ascending=False)
-
-        # Identify most/least similar fold pairs
-        most_pair = None
-        least_pair = None
-        if fold_sim.shape[0] >= 2:
-            # consider only upper triangle (excluding diagonal)
-            sim_vals = fold_sim.where(np.triu(np.ones(fold_sim.shape), k=1).astype(bool))
-            stacked = sim_vals.stack(future_stack=True)  # MultiIndex (seed_i, seed_j) -> similarity
-            if len(stacked) > 0:
-                (i_max, j_max), v_max = stacked.idxmax(), float(stacked.max())
-                (i_min, j_min), v_min = stacked.idxmin(), float(stacked.min())
-                most_pair = (i_max, j_max, v_max)
-                least_pair = (i_min, j_min, v_min)
-
-        out = {
-            "fold_similarity_matrix": fold_sim,
-            "fold_scores": fold_scores,
-            "most_similar_pair": most_pair,
-            "least_similar_pair": least_pair,
-            "fold_method_matrix": M,
-        }
-
-        if return_pairwise and fold_sim.shape[0] >= 2:
-            sim_vals = fold_sim.where(np.triu(np.ones(fold_sim.shape), k=1).astype(bool))
-
-            # Ensure the row/col index level names are unique before stacking,
-            # otherwise reset_index can try to insert a column that already exists.
-            idx_name = fold_sim.index.name or self.seed_column or "fold"
-            col_name = fold_sim.columns.name or self.seed_column or "fold"
-
-            # Use internal unique names for stack/reset_index
-            sim_vals_safe = sim_vals.copy()
-            sim_vals_safe.index = sim_vals_safe.index.rename(f"__{idx_name}_i")
-            sim_vals_safe.columns = sim_vals_safe.columns.rename(f"__{col_name}_j")
-
-            # pandas >= 2.1: support future_stack=True (silences FutureWarning)
-            try:
-                s = sim_vals_safe.stack(future_stack=True)
-            except TypeError:
-                s = sim_vals_safe.stack(dropna=True)
-            s = s.dropna()
-
-            pairwise = s.rename("similarity").reset_index()
-
-            # Rename back to desired output names (avoid collisions on rename too)
-            col_i = f"{self.seed_column}_i"
-            col_j = f"{self.seed_column}_j"
-            pairwise = pairwise.rename(
-                columns={
-                    f"__{idx_name}_i": col_i,
-                    f"__{col_name}_j": col_j,
-                }
-            )
-
-            pairwise = pairwise.sort_values("similarity", ascending=False).reset_index(drop=True)
-            out["pairwise"] = pairwise
-        return out
-
-    def rank_datasets_by_fold_similarity(
-        self,
-        results_per_task: pd.DataFrame,
-        value_col: str = RANK,
-        *,
-        similarity: str = "spearman",  # {"spearman", "pearson"}
-        agg_fold_score: str = "mean_pairwise",  # {"mean_pairwise", "median_pairwise"}
-        min_folds: int = 2,
-        min_methods: int | None = None,
-        include_pairwise_extremes: bool = True,
-        # --- new: stability estimation controls ---
-        target_reliability: float = 0.90,
-        stability_cap_folds: int = 100,
-        stability_conservative: bool = True,
-        stability_rho_floor: float = 0.01,
-    ) -> dict:
-        """Rank datasets by how consistently their folds/seeds agree with each other,
-        and estimate how many folds are needed to get a stable global ordering.
-
-        Assumes `results_per_task` was computed with `include_seed_col=True` so that
-        each row corresponds to (task, seed, method) with a metric like rank/error.
-
-        For each dataset:
-          1) build fold x method matrix
-          2) compute fold-fold similarity matrix (correlation across methods)
-          3) summarize it into a single "fold agreement" score
-          4) estimate stability/reliability of the k-fold aggregate and folds needed
-             to hit `target_reliability`, using Spearman–Brown style extrapolation:
-                Rel(k) = (k*rho) / (1 + (k-1)*rho)
-             where rho ~= fold_agreement.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-            Output of `self.compute_results_per_task(..., include_seed_col=True)`.
-            Must include `self.seed_column`.
-        value_col : str, default RANK
-            Column used to compare method performance across folds.
-        similarity : {"spearman","pearson"}, default "spearman"
-            Similarity metric for fold-fold correlation across methods.
-        agg_fold_score : {"mean_pairwise","median_pairwise"}, default "mean_pairwise"
-            How to aggregate fold-fold similarities into a dataset score.
-        min_folds : int, default 2
-            Require at least this many folds/seeds for a dataset to be scored.
-        min_methods : int | None
-            If set, require each fold have at least this many methods present.
-            Passed through to `dataset_fold_similarity`.
-        include_pairwise_extremes : bool, default True
-            If True, include the most/least similar fold pair per dataset.
-        target_reliability : float, default 0.90
-            Stability threshold τ in (0,1). Higher = stricter stability requirement.
-        stability_cap_folds : int, default 100
-            Maximum folds to return for `folds_needed_for_stability@τ`.
-        stability_conservative : bool, default True
-            If True, treat rho<=0 as "not stably orderable" and return cap (with rho floored).
-        stability_rho_floor : float, default 0.01
-            Minimum rho used when conservative and rho is extremely small/negative.
-
-        Returns:
-        -------
-        out : dict
-            - "dataset_ranking": pd.DataFrame indexed by dataset with columns:
-                ["fold_agreement", "num_folds", "num_methods_min", "num_methods_mean",
-                 "rho_used", "stability_at_num_folds@{τ}", "folds_needed_for_stability@{τ}",
-                 "most_similar_pair", "least_similar_pair" (optional), "error" (optional)]
-              sorted descending by fold_agreement.
-            - "per_dataset": dict[str, dict]
-              Lightweight per-dataset summary.
-            - "stability_params": dict
-              Echo of stability configuration.
-        """
-        if self.seed_column is None:
-            raise ValueError(
-                "BenchmarkEvaluator.seed_column is None, but fold similarity ranking requires a seed/fold column."
-            )
-        if self.seed_column not in results_per_task.columns:
-            raise ValueError(
-                f"results_per_task must include seed column {self.seed_column!r}. "
-                "Ensure you called compute_results_per_task(..., include_seed_col=True).",
-            )
-        if agg_fold_score not in {"mean_pairwise", "median_pairwise"}:
-            raise ValueError(
-                f"agg_fold_score must be 'mean_pairwise' or 'median_pairwise', got {agg_fold_score!r}",
-            )
-        if similarity not in {"spearman", "pearson"}:
-            raise ValueError(f"similarity must be 'spearman' or 'pearson', got {similarity!r}")
-        if not (0 < float(target_reliability) < 1):
-            raise ValueError(f"target_reliability must be in (0,1), got {target_reliability!r}")
-
-        tau = float(target_reliability)
-        tau_tag = f"{tau:.2f}".rstrip("0").rstrip(".")  # e.g. "0.9" or "0.95"
-        col_stability_at_k = "stability_at_num_folds"
-        col_folds_needed = f"folds_needed_for_stability@{tau_tag}"
-
-        def _rel(k: int, rho: float) -> float:
-            if k <= 0 or not np.isfinite(rho) or rho <= 0:
-                return 0.0
-            return (k * rho) / (1.0 + (k - 1) * rho)
-
-        def _folds_needed(rho: float) -> int:
-            # k >= tau(1-rho)/(rho(1-tau))
-            if not np.isfinite(rho) or rho <= 0:
-                return int(stability_cap_folds)
-            k_float = (tau * (1.0 - rho)) / (rho * (1.0 - tau))
-            k_req = int(np.ceil(k_float))
-            return max(1, min(int(stability_cap_folds), k_req))
-
-        datasets = list(pd.Index(results_per_task[self.task_col].unique()).sort_values())
-
-        rows: list[dict] = []
-        per_dataset: dict[str, dict] = {}
-
-        for ds in datasets:
-            try:
-                out_ds = self.dataset_fold_similarity(
-                    results_per_task=results_per_task,
-                    dataset=ds,
-                    value_col=value_col,
-                    similarity=similarity,
-                    agg_across_methods="mean",  # not used for dataset score
-                    min_methods=min_methods,
-                    return_pairwise=False,
-                )
-            except Exception as e:
-                rows.append(
-                    {
-                        self.task_col: ds,
-                        "fold_agreement": np.nan,
-                        "num_folds": 0,
-                        "num_methods_min": np.nan,
-                        "num_methods_mean": np.nan,
-                        "rho_used": np.nan,
-                        col_stability_at_k: np.nan,
-                        col_folds_needed: np.nan,
-                        "error": repr(e),
-                    }
-                )
-                continue
-
-            fold_sim = out_ds["fold_similarity_matrix"]
-            fold_method = out_ds["fold_method_matrix"]
-
-            num_folds = int(fold_sim.shape[0])
-            methods_per_fold = fold_method.notna().sum(axis=1)
-            num_methods_min = int(methods_per_fold.min()) if len(methods_per_fold) else np.nan
-            num_methods_mean = float(methods_per_fold.mean()) if len(methods_per_fold) else np.nan
-
-            if num_folds < min_folds:
-                rows.append(
-                    {
-                        self.task_col: ds,
-                        "fold_agreement": np.nan,
-                        "num_folds": num_folds,
-                        "num_methods_min": num_methods_min,
-                        "num_methods_mean": num_methods_mean,
-                        "rho_used": np.nan,
-                        col_stability_at_k: np.nan,
-                        col_folds_needed: np.nan,
-                        "error": f"Too few folds (min_folds={min_folds})",
-                    }
-                )
-                continue
-
-            # Aggregate off-diagonal similarities into dataset agreement score (rho estimate)
-            sim_vals = fold_sim.to_numpy(copy=True)
-            np.fill_diagonal(sim_vals, np.nan)
-
-            rho = float(np.nanmean(sim_vals)) if agg_fold_score == "mean_pairwise" else float(np.nanmedian(sim_vals))
-
-            rho_used = rho
-            notes = None
-            if stability_conservative:
-                # clamp to [-1,1]
-                if rho_used > 1:
-                    rho_used = 1.0
-                if rho_used < -1:
-                    rho_used = -1.0
-                if rho_used <= 0:
-                    # treat as not reliably aggregatable; use floor for rel(k) curve but folds_needed -> cap
-                    notes = "rho<=0; stability may not be achievable by averaging folds alone"
-                    rho_used = max(float(stability_rho_floor), 0.0)
-
-            stability_at_k = float(_rel(num_folds, rho_used))
-            folds_needed = int(_folds_needed(rho_used))
-
-            row = {
-                self.task_col: ds,
-                "fold_agreement": rho,  # raw estimate from observed fold similarities
-                "num_folds": num_folds,
-                "num_methods_min": num_methods_min,
-                "num_methods_mean": num_methods_mean,
-                "rho_used": rho_used,  # what we used for stability extrapolation
-                col_stability_at_k: stability_at_k,
-                col_folds_needed: folds_needed,
-            }
-            if notes is not None:
-                row["stability_note"] = notes
-
-            if include_pairwise_extremes and num_folds >= 2:
-                tri = np.triu(np.ones_like(sim_vals, dtype=bool), k=1)
-                vals = sim_vals[tri]
-                if np.isfinite(vals).any():
-                    fold_labels = list(fold_sim.index)
-                    ij = np.argwhere(tri)
-                    k_max = int(np.nanargmax(vals))
-                    i_max, j_max = map(int, ij[k_max])
-                    row["most_similar_pair"] = (fold_labels[i_max], fold_labels[j_max], float(vals[k_max]))
-
-                    k_min = int(np.nanargmin(vals))
-                    i_min, j_min = map(int, ij[k_min])
-                    row["least_similar_pair"] = (fold_labels[i_min], fold_labels[j_min], float(vals[k_min]))
-                else:
-                    row["most_similar_pair"] = None
-                    row["least_similar_pair"] = None
-
-            rows.append(row)
-
-            per_dataset[ds] = {
-                "fold_agreement": rho,
-                "num_folds": num_folds,
-                "num_methods_min": num_methods_min,
-                "num_methods_mean": num_methods_mean,
-                "rho_used": rho_used,
-                col_stability_at_k: stability_at_k,
-                col_folds_needed: folds_needed,
-            }
-            if notes is not None:
-                per_dataset[ds]["stability_note"] = notes
-            if include_pairwise_extremes:
-                per_dataset[ds]["most_similar_pair"] = row.get("most_similar_pair")
-                per_dataset[ds]["least_similar_pair"] = row.get("least_similar_pair")
-
-        ranking = pd.DataFrame(rows).set_index(self.task_col)
-        ranking = ranking.sort_values(by="fold_agreement", ascending=False)
-
-        return {
-            "dataset_ranking": ranking,
-            "per_dataset": per_dataset,
-            "stability_params": {
-                "target_reliability": tau,
-                "cap_folds": stability_cap_folds,
-                "conservative": stability_conservative,
-                "rho_floor": stability_rho_floor,
-                "similarity": similarity,
-                "agg_fold_score": agg_fold_score,
-            },
-        }
-
-    def jitter_all_datasets(
-        self,
-        results_per_task: pd.DataFrame,
-        *,
-        rank_col: str = RANK,
-        metric: str = "winrate",
-        datasets: list[str] | None = None,
-        return_per_method: bool = False,
-        sort_by: str = "jitter_mean",
-        ascending: bool = True,
-    ) -> tuple[pd.DataFrame, dict[str, dict] | None]:
-        """Compute fold jitter metrics for each dataset.
-
-        This is a wrapper around `self.dataset_jitter(...)`.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-            Output of `self.compute_results_per_task(..., include_seed_col=True)`.
-        rank_col : str, default RANK
-            Column read for per-fold ranks (rescaled to winrate when metric="winrate").
-        metric : {"winrate", "rank"}, default "winrate"
-            Per-method per-fold score used to compute jitter. ``"winrate"`` rescales the
-            rank to ``1 - (rank - 1)/(M - 1)`` so jitter values are in [0, 1] and
-            comparable across benchmarks with different method counts.
-        datasets : list[str] | None
-            If specified, only compute for these datasets. Otherwise uses all unique datasets.
-        return_per_method : bool, default False
-            If True, also return a dict keyed by dataset with per-method jitter outputs.
-        sort_by : str, default "jitter_mean"
-            Column name to sort the returned DataFrame by.
-        ascending : bool, default True
-            Sort direction.
-
-        Returns:
-        -------
-        df : pd.DataFrame
-            One row per dataset with:
-            ["dataset", "metric", "num_folds", "num_methods", "jitter_mean", "pairwise_jitter_mean"]
-            plus "error" if a dataset fails.
-        per_method : dict[str, dict] | None
-            If return_per_method=True, a dict of per-dataset detailed outputs; else None.
-        """
-        if self.seed_column is None:
-            raise ValueError("seed_column must be set to compute fold jitter.")
-        if self.seed_column not in results_per_task.columns:
-            raise ValueError(
-                f"results_per_task must include seed column {self.seed_column!r}. "
-                "Ensure you called compute_results_per_task(..., include_seed_col=True).",
-            )
-
-        if datasets is None:
-            datasets = list(pd.Index(results_per_task[self.task_col].unique()).sort_values())
-
-        rows: list[dict] = []
-        per_method: dict[str, dict] | None = {} if return_per_method else None
-
-        for ds in datasets:
-            try:
-                out = self.dataset_jitter(
-                    results_per_task=results_per_task,
-                    dataset=ds,
-                    rank_col=rank_col,
-                    metric=metric,
-                    return_per_method=return_per_method,
-                )
-                # ensure consistent row schema
-                rows.append(
-                    {
-                        "dataset": out["dataset"],
-                        "metric": out["metric"],
-                        "num_folds": out["num_folds"],
-                        "num_methods": out["num_methods"],
-                        "jitter_mean": out["jitter_mean"],
-                        "pairwise_jitter_mean": out["pairwise_jitter_mean"],
-                    }
-                )
-
-                if return_per_method:
-                    assert per_method is not None
-                    # Keep only the per-method parts to avoid duplication
-                    per_method[ds] = {
-                        "per_method_jitter_mean": out.get("per_method_jitter_mean", None),
-                        "per_method_pairwise_jitter_mean": out.get("per_method_pairwise_jitter_mean", None),
-                    }
-
-            except Exception as e:
-                rows.append(
-                    {
-                        "dataset": ds,
-                        "metric": metric,
-                        "num_folds": np.nan,
-                        "num_methods": np.nan,
-                        "jitter_mean": np.nan,
-                        "pairwise_jitter_mean": np.nan,
-                        "error": repr(e),
-                    }
-                )
-                if return_per_method:
-                    assert per_method is not None
-                    per_method[ds] = {"error": repr(e)}
-
-        df = pd.DataFrame(rows)
-
-        if sort_by is not None and sort_by in df.columns:
-            df = df.sort_values(by=sort_by, ascending=ascending, na_position="last").reset_index(drop=True)
-
-        return df, per_method
-
-    def dataset_jitter(
-        self,
-        results_per_task: pd.DataFrame,
-        dataset: str,
-        *,
-        rank_col: str = RANK,
-        metric: str = "winrate",
-        return_per_method: bool = False,
-    ) -> dict:
-        """Compute fold-instability metrics for a dataset.
-
-        ``metric`` selects the per-fold per-method score:
-          - ``"winrate"`` (default): per-fold winrate ``1 - (rank - 1) / (M - 1)``,
-            in [0, 1] regardless of M. Comparable across benchmarks with different
-            method counts.
-          - ``"rank"``: raw per-fold rank from ``rank_col``, scale depends on M.
-
-        Metrics returned:
-          - jitter_mean:
-                E_{fold, method}[ |score_fold - score_global| ]
-          - pairwise_jitter_mean:
-                E_{fold1<fold2, method}[ |score_f1 - score_f2| ]
-
-        Assumes results_per_task was computed with include_seed_col=True.
-
-        Parameters
-        ----------
-        results_per_task : pd.DataFrame
-        dataset : str
-        rank_col : str
-            Source rank column (default: RANK).
-        metric : {"winrate", "rank"}, default "winrate"
-        return_per_method : bool
-            If True, also return per-method jitter statistics.
-
-        Returns:
-        -------
-        dict
-        """
-        if metric not in ("winrate", "rank"):
-            raise ValueError(f"metric must be 'winrate' or 'rank', got {metric!r}")
-        if self.seed_column is None:
-            raise ValueError("seed_column must be set to compute fold jitter.")
-        if self.seed_column not in results_per_task.columns:
-            raise ValueError(
-                "results_per_task must include seed column. Call compute_results_per_task(..., include_seed_col=True)",
-            )
-
-        df = results_per_task.loc[results_per_task[self.task_col] == dataset].copy()
-
-        if df.empty:
-            raise ValueError(f"No rows found for dataset {dataset!r}")
-
-        # Pivot: folds x methods (values are ranks; rescaled to winrate below if requested)
-        M = df.pivot(
-            index=self.seed_column,
-            columns=self.method_col,
-            values=rank_col,
-        )
-
-        # Drop methods with any missing folds
-        M = M.dropna(axis=1)
-
-        folds = M.index.tolist()
-        methods = M.columns.tolist()
-
-        num_folds = len(folds)
-        num_methods = len(methods)
-
-        if num_folds < 2:
-            raise ValueError("Need at least 2 folds to compute jitter.")
-        if num_methods < 2:
-            raise ValueError("Need at least 2 methods to compute jitter.")
-
-        if metric == "winrate":
-            # winrate = 1 - (rank - 1) / (M - 1); in [0, 1] regardless of M.
-            M = 1.0 - (M - 1.0) / (num_methods - 1)
-
-        # ---------------------------------------------------------
-        # 1) Expected jitter: deviation of per-fold score from the global mean
-        # ---------------------------------------------------------
-
-        global_score = M.mean(axis=0)
-
-        abs_dev = (M - global_score).abs()
-
-        jitter_mean = abs_dev.values.mean()
-
-        # ---------------------------------------------------------
-        # 2) Pairwise jitter: average absolute change between folds
-        # ---------------------------------------------------------
-
-        pairwise_changes = []
-
-        for i in range(num_folds):
-            for j in range(i + 1, num_folds):
-                diff = (M.iloc[i] - M.iloc[j]).abs()
-                pairwise_changes.append(diff.values.mean())
-
-        pairwise_jitter_mean = float(np.mean(pairwise_changes))
-
-        result = {
-            "dataset": dataset,
-            "metric": metric,
-            "num_folds": num_folds,
-            "num_methods": num_methods,
-            "jitter_mean": float(jitter_mean),
-            "pairwise_jitter_mean": pairwise_jitter_mean,
-        }
-
-        if return_per_method:
-            result["per_method_jitter_mean"] = abs_dev.mean(axis=0)
-            result["per_method_pairwise_jitter_mean"] = pd.concat(
-                [(M.iloc[i] - M.iloc[j]).abs() for i in range(num_folds) for j in range(i + 1, num_folds)],
-                axis=1,
-            ).mean(axis=1)
-
-        return result
-
-    def dataset_jitter_bootstrap_curve(
-        self,
-        results_per_task: pd.DataFrame,
-        dataset: str,
-        *,
-        rank_col: str = RANK,
-        metric: str = "winrate",
-        k_values: list[int] | None = None,
-        n_bootstrap: int = 500,
-        seed: int = 0,
-        drop_methods_with_any_missing: bool = True,
-    ) -> pd.DataFrame:
-        """Bootstrap how the k-fold *mean per-method score* converges to the global mean.
-
-        For each bootstrap replicate:
-          - sample k folds WITH replacement
-          - compute per-method mean score across those k folds  (mu_hat_k)
-          - compare to per-method global mean score across ALL folds (mu_global)
-          - jitter_k = mean_m |mu_hat_k[m] - mu_global[m]|
-
-        ``metric`` selects the per-fold per-method score:
-          - ``"winrate"`` (default): per-fold winrate ``1 - (rank - 1) / (M - 1)``,
-            in [0, 1] regardless of M. Comparable across datasets/benchmarks with
-            different method counts.
-          - ``"rank"``: raw per-fold rank from ``rank_col``, scale depends on M.
-
-        Returns a DataFrame with mean/median/95% CI of jitter_k vs k, plus a
-        ``metric`` column identifying which score was used.
-        """
-        if metric not in ("winrate", "rank"):
-            raise ValueError(f"metric must be 'winrate' or 'rank', got {metric!r}")
-        if self.seed_column is None:
-            raise ValueError("seed_column must be set to compute fold jitter.")
-        if self.seed_column not in results_per_task.columns:
-            raise ValueError(
-                f"results_per_task must include seed column {self.seed_column!r}. "
-                "Ensure you called compute_results_per_task(..., include_seed_col=True).",
-            )
-
-        df = results_per_task.loc[results_per_task[self.task_col] == dataset].copy()
-        if df.empty:
-            raise ValueError(f"No rows found for dataset {dataset!r}")
-
-        # folds x methods matrix of ranks (the canonical per-fold score; winrate is a
-        # per-fold linear rescale of rank, see below)
-        M_df = df.pivot(index=self.seed_column, columns=self.method_col, values=rank_col)
-
-        if drop_methods_with_any_missing:
-            M_df = M_df.dropna(axis=1)
-        elif M_df.isna().any().any():
-            raise ValueError(
-                "Found NaNs in fold×method matrix. Set drop_methods_with_any_missing=True "
-                "or impute missing values before bootstrapping.",
-            )
-
-        M_full = M_df.to_numpy(dtype=float)  # shape (F, M), values are ranks
-        F, M = M_full.shape
-        if F < 1:
-            raise ValueError("No folds available after filtering.")
-        if M < 2:
-            raise ValueError("Need at least 2 methods after filtering.")
-
-        if metric == "winrate":
-            # winrate = 1 - (rank - 1) / (M - 1); in [0, 1] regardless of M.
-            # Equivalent to the pairwise winrate convention with average-rank tie handling.
-            M_full = 1.0 - (M_full - 1.0) / (M - 1.0)
-
-        # Global per-method mean score across ALL folds (fixed reference)
-        mu_global = M_full.mean(axis=0)  # shape (M,)
-
-        if k_values is None:
-            k_values = list(range(1, F + 1))
-        else:
-            k_values = [int(k) for k in k_values]
-            if any(k < 1 for k in k_values):
-                raise ValueError("All k_values must be >= 1.")
-            if any(k > F for k in k_values):
-                raise ValueError(f"All k_values must be <= number of available folds ({F}).")
-
-        rng = np.random.default_rng(seed)
-
-        def _summ(x: np.ndarray) -> dict:
-            return dict(
-                mean=float(np.mean(x)),
-                p50=float(np.quantile(x, 0.50)),
-                p025=float(np.quantile(x, 0.025)),
-                p975=float(np.quantile(x, 0.975)),
-            )
-
-        rows: list[dict] = []
-
-        for k in k_values:
-            jit = np.empty(n_bootstrap, dtype=float)
-
-            for b in range(n_bootstrap):
-                # Sample folds WITH replacement: classical bootstrap of the k-fold
-                # mean against the all-folds reference. At k=F this still produces
-                # non-zero jitter (Monte-Carlo resampling noise of the all-folds
-                # mean estimate) — without that, Φ=1 collapses to 0 in the strip.
-                idx = rng.choice(F, size=k, replace=True)
-                M_sel = M_full[idx, :]  # (k, M)
-
-                mu_hat = M_sel.mean(axis=0)  # per-method mean score across sampled folds
-                jit[b] = float(np.abs(mu_hat - mu_global).mean())
-
-            s = _summ(jit)
-            rows.append(
-                {
-                    "dataset": dataset,
-                    "metric": metric,
-                    "k": k,
-                    "n_bootstrap": n_bootstrap,
-                    "n_folds_available": F,
-                    "n_methods": M,
-                    "jitter_mean": s["mean"],
-                    "jitter_p50": s["p50"],
-                    "jitter_p025": s["p025"],
-                    "jitter_p975": s["p975"],
-                }
-            )
-
-        return pd.DataFrame(rows).sort_values("k").reset_index(drop=True)
-
-    def jitter_bootstrap_curve_all_datasets(
-        self,
-        results_per_task: pd.DataFrame,
-        *,
-        rank_col: str = RANK,
-        metric: str = "winrate",
-        k_values: list[int] | None = None,
-        n_bootstrap: int = 300,
-        seed: int = 0,
-        drop_methods_with_any_missing: bool = True,
-    ) -> pd.DataFrame:
-        datasets = list(pd.Index(results_per_task[self.task_col].unique()).sort_values())
-        dfs: list[pd.DataFrame] = []
-
-        for ds in datasets:
-            try:
-                d = self.dataset_jitter_bootstrap_curve(
-                    results_per_task=results_per_task,
-                    dataset=ds,
-                    rank_col=rank_col,
-                    metric=metric,
-                    k_values=k_values,
-                    n_bootstrap=n_bootstrap,
-                    seed=seed,
-                    drop_methods_with_any_missing=drop_methods_with_any_missing,
-                )
-                dfs.append(d)
-            except Exception as e:
-                dfs.append(
-                    pd.DataFrame(
-                        [
-                            {
-                                "dataset": ds,
-                                "metric": metric,
-                                "k": np.nan,
-                                "n_bootstrap": n_bootstrap,
-                                "error": repr(e),
-                            }
-                        ]
-                    )
-                )
-
-        return pd.concat(dfs, axis=0, ignore_index=True)
-
-    @staticmethod
-    def estimate_folds_for_stable_ordering(
-        fold_agreement: float,
-        num_folds: int,
-        *,
-        target_reliability: float = 0.90,  # "stable ordering" threshold
-        cap: int = 100,  # safety cap
-        conservative: bool = True,
-        rho_floor: float = 0.01,
-    ) -> dict:
-        """Estimate how many folds are needed to get a stable global ordering on a dataset,
-        based on observed inter-fold agreement.
-
-        Uses Spearman–Brown style extrapolation:
-            Rel(k) = (k*rho) / (1 + (k-1)*rho)
-        where rho ~ mean pairwise fold correlation ("fold_agreement").
-
-        Parameters
-        ----------
-        fold_agreement : float
-            Mean pairwise fold similarity (e.g., Spearman correlation across methods).
-        num_folds : int
-            Number of folds used to estimate fold_agreement.
-            Used mainly for reporting; the extrapolation itself uses fold_agreement.
-        target_reliability : float, default 0.90
-            Target reliability for considering the ordering "stable".
-        cap : int, default 100
-            Maximum folds to return (avoid absurd numbers).
-        conservative : bool, default True
-            If True, clamp rho into a plausible range and warn on edge cases.
-        rho_floor : float, default 0.01
-            Minimum rho to use when conservative and rho is extremely small/negative.
-
-        Returns:
-        -------
-        dict with:
-            - "rho": used rho
-            - "target_reliability": tau
-            - "k_required": estimated folds needed (int, capped)
-            - "rel_at_num_folds": estimated reliability at the observed num_folds
-            - "rel_curve": optional small table (k, Rel(k)) for k in [1..min(cap, 20)]
-            - "notes": list[str]
-        """
-        notes = []
-        rho = float(fold_agreement)
-
-        if not np.isfinite(rho):
-            raise ValueError("fold_agreement must be finite.")
-
-        # Correlation should be in [-1, 1]; we only have meaningful fold-averaging reliability for rho > 0
-        if conservative:
-            if rho > 1:
-                notes.append("rho > 1 encountered; clamping to 1.")
-                rho = 1.0
-            if rho < -1:
-                notes.append("rho < -1 encountered; clamping to -1.")
-                rho = -1.0
-
-            if rho <= 0:
-                notes.append(
-                    "fold_agreement <= 0 suggests folds disagree or are unrelated; "
-                    "a stable global ordering may not be achievable by averaging folds alone. "
-                    "Returning cap.",
-                )
-                rho = max(rho_floor, 0.0)
-
-        tau = float(target_reliability)
-        if not (0 < tau < 1):
-            raise ValueError("target_reliability must be in (0, 1).")
-
-        def rel(k: int) -> float:
-            if k <= 0:
-                return np.nan
-            if rho <= 0:
-                return 0.0
-            return (k * rho) / (1.0 + (k - 1) * rho)
-
-        rel_at_num = rel(int(num_folds))
-
-        # Solve k >= tau(1-rho)/(rho(1-tau))
-        if rho <= 0:
-            k_req = cap
-        else:
-            k_float = (tau * (1.0 - rho)) / (rho * (1.0 - tau))
-            k_req = int(np.ceil(k_float))
-            k_req = max(1, k_req)
-
-        if k_req > cap:
-            notes.append(f"Estimated folds ({k_req}) exceed cap ({cap}); returning cap.")
-            k_req = cap
-
-        rel_curve = [(k, rel(k)) for k in range(1, min(cap, 20) + 1)]
-        rel_curve_df = pd.DataFrame(rel_curve, columns=["k", "reliability"])
-
-        return {
-            "rho": rho,
-            "target_reliability": tau,
-            "k_required": k_req,
-            "rel_at_num_folds": rel_at_num,
-            "rel_curve": rel_curve_df,
-            "notes": notes,
-        }
 
 
 def get_bootstrap_result_lst(
